+++ src/renderer/features/fellowship-referendum-details/service.ts
@@ -1,27 +1,4 @@
-import { type Track } from "@/domains/collectives";
-
-interface RankActivityThreshold {
-  activity: number | null;
-  agreement: number | null;
-}
-
-const rankThresholds: Record<number, RankActivityThreshold> = {
-  0: { activity: null, agreement: null },
-  1: { activity: 90, agreement: null },
-  2: { activity: 80, agreement: null },
-  3: { activity: 70, agreement: 100 },
-  4: { activity: 60, agreement: 90 },
-  5: { activity: 50, agreement: 80 },
-  6: { activity: 40, agreement: 70 },
-  7: { activity: null, agreement: null },
-  8: { activity: null, agreement: null },
-  9: { activity: null, agreement: null },
-};
-
-function getActivityAndAgreementThresholds(rank: number) {
-  return rankThresholds[rank] ?? { activity: null, agreement: null };
-}
-
+import { type Track } from '@/domains/collectives';
 const getRankTitle = (rank: number, relatedTrack: Track[] | null | undefined) => {
   const name = relatedTrack?.find(t => t.id === rank)?.name;
 
@@ -31,6 +8,5 @@ const getRankTitle = (rank: number, relatedTrack: Track[] | null | undefined) =>
 };
 
 export const detailsService = {
-  getActivityAndAgreementThresholds,
-  getRankTitle
+  getRankTitle,
 };
